Refactoring results download code in athena query tests
