+++ integration-tests/athenaQuery.spec.ts
@@ -25,7 +25,7 @@ import {
   dynamoDBItemPIITypesOnly
 } from './constants/dynamoDBItemDetails'
 import { deleteAuditDataWithPrefix } from './utils/aws/s3DeleteAuditDataWithPrefix'
-// import * as CSV from 'csv-string'
+import { downloadResultsFileAndParseData } from './utils/queryResults/downloadAndParseResults'
 
 describe('Athena Query SQL generation and execution', () => {
   jest.setTimeout(90000)
@@ -83,29 +83,11 @@ describe('Athena Query SQL generation and execution', () => {
       const value = await getValueFromDynamoDB(randomTicketId, 'athenaQueryId')
       expect(value?.athenaQueryId.S).toBeDefined()
 
-      //TODO: REPLACE with: call link within email to download results
-
-      /*const downloadHash = await waitForDownloadHash(randomTicketId)
-      expect(downloadHash).toBeDefined()
-      console.log('Download Hash: ' + downloadHash)
-
-      const downloadPageHTML = await getSecureDownloadPageHTML(downloadHash)
-      expect(downloadHash.startsWith('<html>')).toBeTrue
-
-      const linkToS3ResultsFile = retrieveS3LinkFromHtml(downloadPageHTML)
-      expect(linkToS3ResultsFile.startsWith('https')).toBeTrue
-
-      const csvData = await downloadResultsCSVFromLink(linkToS3ResultsFile)
-      console.log('CSV data: ' + csvData)
-      const csvRows = CSV.parse(csvData, { output: 'objects' })
-      console.log(csvRows)
+      const csvRows = await downloadResultsFileAndParseData(randomTicketId)
 
       expect(csvRows.length).toEqual(1)
       expect(csvRows[0].birthdate_value).toEqual(EXPECTED_RESULTS_BIRTHDATE)
-      expect(csvRows[0].address_buildingname).toEqual(EXPECTED_BUILDING_NAME)*/
-
-      console.log(EXPECTED_RESULTS_BIRTHDATE)
-      console.log(EXPECTED_BUILDING_NAME)
+      expect(csvRows[0].address_buildingname).toEqual(EXPECTED_BUILDING_NAME)
     })
 
     it('Successful Athena processing - requests having only PII type', async () => {
@@ -137,29 +119,11 @@ describe('Athena Query SQL generation and execution', () => {
       const value = await getValueFromDynamoDB(randomTicketId, 'athenaQueryId')
       expect(value?.athenaQueryId.S).toBeDefined()
 
-      //TODO: REPLACE with: call link within email to download results
-
-      /*const downloadHash = await waitForDownloadHash(randomTicketId)
-      expect(downloadHash).toBeDefined()
-      console.log('Download Hash: ' + downloadHash)
-
-      const downloadPageHTML = await getSecureDownloadPageHTML(downloadHash)
-      expect(downloadHash.startsWith('<html>')).toBeTrue
-
-      const linkToS3ResultsFile = retrieveS3LinkFromHtml(downloadPageHTML)
-      expect(linkToS3ResultsFile.startsWith('https')).toBeTrue
-
-      const csvData = await downloadResultsCSVFromLink(linkToS3ResultsFile)
-      console.log('CSV data: ' + csvData)
-      const csvRows = CSV.parse(csvData, { output: 'objects' })
-      console.log(csvRows)
+      const csvRows = await downloadResultsFileAndParseData(randomTicketId)
 
       expect(csvRows.length).toEqual(1)
       expect(csvRows[0].name).toEqual(EXPECTED_NAME)
-      expect(csvRows[0].addresses).toEqual(EXPECTED_ADDRESSES)*/
-
-      console.log(EXPECTED_NAME)
-      console.log(EXPECTED_ADDRESSES)
+      expect(csvRows[0].addresses).toEqual(EXPECTED_ADDRESSES)
     })
 
     it('Successful Athena processing - requests having both data paths and PII types', async () => {
@@ -193,31 +157,13 @@ describe('Athena Query SQL generation and execution', () => {
       const value = await getValueFromDynamoDB(randomTicketId, 'athenaQueryId')
       expect(value?.athenaQueryId.S).toBeDefined()
 
-      /*const downloadHash = await waitForDownloadHash(randomTicketId)
-      expect(downloadHash).toBeDefined()
-      console.log('Download Hash: ' + downloadHash)
-
-      const downloadPageHTML = await getSecureDownloadPageHTML(downloadHash)
-      expect(downloadHash.startsWith('<html>')).toBeTrue
-
-      const linkToS3ResultsFile = retrieveS3LinkFromHtml(downloadPageHTML)
-      expect(linkToS3ResultsFile.startsWith('https')).toBeTrue
-
-      const csvData = await downloadResultsCSVFromLink(linkToS3ResultsFile)
-      console.log('CSV data: ' + csvData)
-      const csvRows = CSV.parse(csvData, { output: 'objects' })
-      console.log(csvRows)
+      const csvRows = await downloadResultsFileAndParseData(randomTicketId)
 
       expect(csvRows.length).toEqual(1)
       expect(csvRows[0].birthdate_value).toEqual(EXPECTED_RESULTS_BIRTHDATE)
       expect(csvRows[0].address_buildingname).toEqual(EXPECTED_BUILDING_NAME)
       expect(csvRows[0].name).toEqual(EXPECTED_NAME)
-      expect(csvRows[0].addresses).toEqual(EXPECTED_ADDRESSES)*/
-
-      console.log(EXPECTED_NAME)
-      console.log(EXPECTED_ADDRESSES)
-      console.log(EXPECTED_BUILDING_NAME)
-      console.log(EXPECTED_RESULTS_BIRTHDATE)
+      expect(csvRows[0].addresses).toEqual(EXPECTED_ADDRESSES)
     })
   })
 
@@ -255,4 +201,29 @@ describe('Athena Query SQL generation and execution', () => {
       assertEventPresent(athenaQueryEvents, ATHENA_HANDLER_INVOKE_ERROR)
     })
   })
+
+  // async function downloadResultsFileAndParseData(ticketId: string): Promise<
+  //   {
+  //     [k: string]: string
+  //   }[]
+  // > {
+  //   const secureDownloadPageUrl = await waitForDownloadUrlFromNotifyEmail(
+  //     ticketId
+  //   )
+  //   expect(secureDownloadPageUrl.startsWith('https')).toBeTrue
+
+  //   const secureDownloadPageHTML = await getSecureDownloadPageHTML(
+  //     secureDownloadPageUrl
+  //   )
+
+  //   const resultsFileS3Link = retrieveS3LinkFromHtml(secureDownloadPageHTML)
+  //   expect(resultsFileS3Link.startsWith('https')).toBeTrue
+
+  //   const csvData = await downloadResultsCSVFromLink(resultsFileS3Link)
+  //   console.log(csvData)
+
+  //   const csvRows = CSV.parse(csvData, { output: 'objects' })
+  //   console.log(csvRows)
+  //   return csvRows
+  // }
 })
